Allow overriding keybindings from config.
Disclaimer:

I have 0 rust experience so I ended up using AI to generate the code.

Notes:

This chagne is config only. It doesn't add any UI to enable changing
the key bindings.  But this is usable as is, and UI is something that
could be added later in a different change.

My personal goal was to enable re-mapping "Shift + Insert" to "Paste" from
"PastePrimary", without having to maintain a fork.

The config file format should be similar to what comsic-settings and
cosmic-comp use for their shortcuts, except that it uses KeyBinding
from libcosmic as the map key and comsic-term's Action enum (instead
of the Binding type and Action enum cosmic-settings uses).  But this
should make the config familar to anyone familar with cosmic-settings
shortcuts.

Here's a sample config:

```
// Example custom keybindings for cosmic-term.
// Save this file as ~/.config/cosmic/com.system76.CosmicTerm/v1/keybindings
{
    // Use Shift+Insert for Paste instead of PastePrimary.
    (modifiers: [Shift], key: "Insert"): Paste,
}
```

Key bindings can alos be disabled, such as

```
{
    (modifiers: [Shift], key: "Insert"): Disable,
}
```

I wasn't sure if it made sense to push modifications into libcomsic
(which would have increased the scope of the change), so I didn't.

KeyBind and Modifiers from libcomsic are not serializable, so I used
`KeyBindRon` and `ModifierLabel` to implement the serialization /
deserialization.

If you think it would be better to just make the libcosmic types
serializable, let me know, and I can try to hack that up.

Testing:

There's some super basic unit tests. I also manually tested both
overriding and disabling actions, as well as modfiying the config
while the terminal is working, and that all seems to work correctly.